fix: resolve type mismatch in PlanService and Test (String to Long)

+++ src/main/kotlin/com/example/planmanager/plan/service/PlanService.kt
@@ -22,7 +22,7 @@ class PlanService(
 
     fun createPlan(userId: String, request: PlanCreateRequest): PlanResponse {
         val plan = PlanEntity(
-            userId = userId,
+            userId = userId.toLong(),
             title = request.title,
             planDate = request.planDate
         )
